Django Api/django_rest/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Deprecated==1.2.13
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
idna==3.4
jwcrypto==1.4.2
mysqlclient==2.1.1
oauthlib==3.2.2
pycparser==2.21
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
wrapt==1.14.1
